streaming/multiple_shards/requirements.txt
Faker==26.0.0
kafka-python==2.0.2